wawmomo backend: configuration des images, continuation

+++ restaurants/wawmomo/back-end/src/imagesConfig/fsConfig.js
@@ -4,6 +4,10 @@ import fs from 'fs';
 
 export const fsDeleteImage = (req, res, next) => {
     try {
+        if (!req.file) {
+            return res.status(400).json({ status: 400, message: "file not found" })
+        }
+        // Dossier pour les images redimensionnées
         const filePath = req.file.path;
         fs.unlink(filePath, (err) => {
             if (err) {
@@ -11,13 +15,11 @@ export const fsDeleteImage = (req, res, next) => {
                 return res.status(500).json({ status: 500, message: "Error deleting file" });
             }
             console.log("File deleted successfully");
-            req.body.res.status = 200
-            req.body.res.message = "File deleted"
             next()
         });
     } catch (error) {
         console.log("deleteImage -> error");
         console.log(error);
-        res.status(500).json({ status: 500, message: "server problem, impossible to delete" })
+        res.status(500).json({ status: 500, message: "server problem, impossible to delete tampon" })
     }
 }
